# swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  OpendbHackAddress:
    description: OpendbHackAddress media type (default view)
    example:
      city: Itaque nisi possimus aut vel vel.
      country: Perferendis pariatur ut non non quia rerum.
      houseNumber: Ut architecto est commodi qui dolorum dignissimos.
      street: Illum eos consectetur ducimus.
      type: Quod sunt quia.
      zip: Reiciendis non quae velit iste impedit debitis.
    properties:
      city:
        description: City,
        example: Itaque nisi possimus aut vel vel.
        type: string
      country:
        description: Country Code,
        example: Perferendis pariatur ut non non quia rerum.
        type: string
      houseNumber:
        description: House number,
        example: Ut architecto est commodi qui dolorum dignissimos.
        type: string
      street:
        description: Street,
        example: Illum eos consectetur ducimus.
        type: string
      type:
        description: Type of address = [„MAILING_ADDRESS“ or „REGISTRATION_ADDRESS“]
        example: Quod sunt quia.
        type: string
      zip:
        description: Zip code,
        example: Reiciendis non quae velit iste impedit debitis.
        type: string
    required:
    - street
    - houseNumber
    - zip
    - city
    - country
    - type
    title: 'Mediatype identifier: application/vnd.opendb.hack.address+json; view=default'
    type: object
  OpendbHackAddressCollection:
    description: OpendbHackAddressCollection is the media type for an array of OpendbHackAddress
      (default view)
    example:
    - city: Itaque nisi possimus aut vel vel.
      country: Perferendis pariatur ut non non quia rerum.
      houseNumber: Ut architecto est commodi qui dolorum dignissimos.
      street: Illum eos consectetur ducimus.
      type: Quod sunt quia.
      zip: Reiciendis non quae velit iste impedit debitis.
    - city: Itaque nisi possimus aut vel vel.
      country: Perferendis pariatur ut non non quia rerum.
      houseNumber: Ut architecto est commodi qui dolorum dignissimos.
      street: Illum eos consectetur ducimus.
      type: Quod sunt quia.
      zip: Reiciendis non quae velit iste impedit debitis.
    - city: Itaque nisi possimus aut vel vel.
      country: Perferendis pariatur ut non non quia rerum.
      houseNumber: Ut architecto est commodi qui dolorum dignissimos.
      street: Illum eos consectetur ducimus.
      type: Quod sunt quia.
      zip: Reiciendis non quae velit iste impedit debitis.
    items:
      $ref: '#/definitions/OpendbHackAddress'
    title: 'Mediatype identifier: application/vnd.opendb.hack.address+json; type=collection'
    type: array
  OpendbHackCashaccount:
    description: OpendbHackCashaccount media type (default view)
    example:
      balance: 0.6645298
      iban: Ullam harum similique ut.
      productDescription: Dolores iusto ipsam eum sint ut.
    properties:
      balance:
        description: Booked balance in EUR
        example: 0.6645298
        format: double
        type: number
      iban:
        description: IBAN of the cash account
        example: Ullam harum similique ut.
        type: string
      productDescription:
        description: Description of the product
        example: Dolores iusto ipsam eum sint ut.
        type: string
    required:
    - iban
    - balance
    - productDescription
    title: 'Mediatype identifier: application/vnd.opendb.hack.cashaccount+json; view=default'
    type: object
  OpendbHackCashaccountCollection:
    description: OpendbHackCashaccountCollection is the media type for an array of
      OpendbHackCashaccount (default view)
    example:
    - balance: 0.6645298
      iban: Ullam harum similique ut.
      productDescription: Dolores iusto ipsam eum sint ut.
    - balance: 0.6645298
      iban: Ullam harum similique ut.
      productDescription: Dolores iusto ipsam eum sint ut.
    items:
      $ref: '#/definitions/OpendbHackCashaccount'
    title: 'Mediatype identifier: application/vnd.opendb.hack.cashaccount+json; type=collection'
    type: array
  OpendbHackTransaction:
    description: OpendbHackTransaction media type (default view)
    example:
      amount: 0.0378779
      counterPartyIban: Sed dolores reiciendis.
      counterPartyName: Autem tenetur laudantium atque quisquam.
      date: Dolor repellat laborum debitis.
      usage: Doloribus unde magnam ea.
    properties:
      amount:
        description: Amount of the transaction. If the amount is positive, the customer
          received money, if the amount is negative the customer lost money,
        example: 0.0378779
        format: double
        type: number
      counterPartyIban:
        description: IBAN of the counter party
        example: Sed dolores reiciendis.
        type: string
      counterPartyName:
        description: Name of the counter party
        example: Autem tenetur laudantium atque quisquam.
        type: string
      date:
        description: Posting date in the format YYYY-MM-DD
        example: Dolor repellat laborum debitis.
        type: string
      usage:
        description: Payment reference
        example: Doloribus unde magnam ea.
        type: string
    required:
    - amount
    - date
    title: 'Mediatype identifier: application/vnd.opendb.hack.transaction+json; view=default'
    type: object
  OpendbHackTransactionCollection:
    description: OpendbHackTransactionCollection is the media type for an array of
      OpendbHackTransaction (default view)
    example:
    - amount: 0.0378779
      counterPartyIban: Sed dolores reiciendis.
      counterPartyName: Autem tenetur laudantium atque quisquam.
      date: Dolor repellat laborum debitis.
      usage: Doloribus unde magnam ea.
    items:
      $ref: '#/definitions/OpendbHackTransaction'
    title: 'Mediatype identifier: application/vnd.opendb.hack.transaction+json; type=collection'
    type: array
  OpendbHackUserinfo:
    description: OpendbHackUserinfo media type (default view)
    example:
      dateOfBirth: Incidunt laborum aperiam numquam enim rerum.
      firstName: Recusandae suscipit enim.
      gender: Qui assumenda.
      lastName: Atque voluptates dolores sit accusantium aut.
    properties:
      dateOfBirth:
        description: Birth date of the user in the format YYYY-MM-DD
        example: Incidunt laborum aperiam numquam enim rerum.
        type: string
      firstName:
        description: First name of the user
        example: Recusandae suscipit enim.
        type: string
      gender:
        description: Gender of the user = [„MALE“ or „FEMALE“ or „UNKOWN“]
        example: Qui assumenda.
        type: string
      lastName:
        description: Last name of the user
        example: Atque voluptates dolores sit accusantium aut.
        type: string
    required:
    - firstName
    - lastName
    - dateOfBirth
    - gender
    title: 'Mediatype identifier: application/vnd.opendb.hack.userinfo+json'
    type: object
  PostTransactionsTransactionsPayload:
    example:
      amount: Odit aut odit tempora.
      creditorBIC: Et est.
      creditorIBAN: Harum est totam voluptatem.
      creditorName: Unde alias.
      currency: Voluptatem ipsa dolorum rem libero.
      debtorBIC: Molestiae repellendus minus iure.
      debtorIBAN: Eius voluptatem consequuntur reiciendis voluptatibus veritatis.
      debtorName: Debitis totam possimus et distinctio.
      remittanceInformation: Sit pariatur reiciendis inventore praesentium.
    properties:
      amount:
        example: Odit aut odit tempora.
        type: string
      creditorBIC:
        example: Et est.
        type: string
      creditorIBAN:
        example: Harum est totam voluptatem.
        type: string
      creditorName:
        example: Unde alias.
        type: string
      currency:
        example: Voluptatem ipsa dolorum rem libero.
        type: string
      debtorBIC:
        example: Molestiae repellendus minus iure.
        type: string
      debtorIBAN:
        example: Eius voluptatem consequuntur reiciendis voluptatibus veritatis.
        type: string
      debtorName:
        example: Debitis totam possimus et distinctio.
        type: string
      remittanceInformation:
        example: Sit pariatur reiciendis inventore praesentium.
        type: string
    required:
    - amount
    - creditorIBAN
    - creditorBIC
    - debtorIBAN
    - debtorBIC
    - currency
    - remittanceInformation
    title: PostTransactionsTransactionsPayload
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error'
    type: object
host: localhost:8880
info:
  description: 'Service to proxy '
  title: The mock for opendb
  version: ""
paths:
  /addresses:
    get:
      description: Get adresses
      operationId: user#GetAddresses
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OpendbHackAddressCollection'
        "201":
          description: Created
      schemes:
      - http
      summary: GetAddresses user
      tags:
      - user
  /cashAccounts:
    get:
      description: Cash accuonts
      operationId: accounts#GetCashAccounts
      parameters:
      - in: query
        name: iban
        required: false
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OpendbHackCashaccountCollection'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      security:
      - token: []
      summary: GetCashAccounts accounts
      tags:
      - accounts
  /transactions:
    get:
      description: Get all transactions
      operationId: transactions#GetTransactions
      parameters:
      - in: query
        name: iban
        required: false
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OpendbHackTransactionCollection'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      security:
      - token: []
      summary: GetTransactions transactions
      tags:
      - transactions
    post:
      description: Create a transaction
      operationId: transactions#PostTransactions
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/PostTransactionsTransactionsPayload'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      security:
      - token: []
      summary: PostTransactions transactions
      tags:
      - transactions
  /userInfo:
    get:
      description: Get User Info
      operationId: user#GetUserInfo
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OpendbHackUserinfo'
      schemes:
      - http
      summary: GetUserInfo user
      tags:
      - user
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
responses:
  Created:
    description: Created
  Unauthorized:
    description: Unauthorized
schemes:
- http
securityDefinitions:
  token:
    in: header
    name: token
    type: apiKey
swagger: "2.0"